refactor Metrics components for new folder structure
